Update script due to review notes
